feat: reorganização de pasta e arquivos para melhor visualização
