--- src/App.css
@@ -1,136 +1,0 @@
-*{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-:root{
-  /* color ligh */
-  --color-font-primary: rgb(0, 0, 0);
-  --color-background-primary: rgb(0, 0, 0);
-  --color-font-secondary: rgb(255, 255, 255);
-  --color-background-secondary: rgb(255, 255, 255);
-  --color-font-menu: rgb(49, 49, 49);
-}
-.App {
-  text-align: left;
-}
-.cabecalho{
-  position: fixed;
-  display: flex;
-}
-
-.menu-below-700{
-  padding: 15px;
-}
-.menu ul{
-  max-width: 90%;
-  display: inline;
-  list-style: none;
-}
-.menu ul li a{
-  font-size: 20px;
-  transition: transform 200ms;
-  transition: background .2s;
-  padding: 15px;
-  display: block;
-  text-decoration: none;
-}
-.menu ul li a:hover{
-  transform: scale(1.02);
-}
-.body {
-  box-sizing:border-box;
-  height: 100vh;
-  font-size: calc(10px + 2vmin);
-  overflow-y: scroll;
-  scrollbar-width: none;
-}
-.body::-webkit-scrollbar {
-  display: none;
-}
-.stats{
-  padding-top: 60px;
-}
-.imgs{
-  text-align: center;
-  padding: 40px auto;
-}
-.imgs img, #stats{
-  display: inline;
-  padding: auto;
-}
-.menu-below-700:hover{
-  cursor: pointer;  
-}
-
-@media (min-width: 700px) {
-  .menu-below-700{
-    display: none;
-  }
-  .cabecalho{
-    width: 100%;
-  }
-  .imgs img{
-    height: 500px;
-    width: 600px;
-  }
-  .menu ul li{
-    display: inline-block;
-  }
-  
-  .menu{
-    margin: auto;
-  }
-}
-
-
-
-/* display between width 1100 - 900  */
-@media (max-width: 1100px) {
-  .menu ul li a{
-    font-size: 15px;
-  }
-}
-@media (max-width: 900px) {
-  .menu ul li a{
-    padding: 10px;
-    font-size: 13px;
-
-  }
-}
-
-
-
-
-/* mobile responsivity 700 */
-@media (max-width: 700px) {
-  .effect .none .inline .open{
-    transition: margin .5s;
-  }
-  .menu-below-700{
-    padding: 10px;
-    max-height: 45px;
-    
-  }.t{
-    width: 100%;
-
-  }.m{
-  }
-  .body {
-    display: block;
-    font-size: calc(10px + 1vmin);
-    overflow-x: hidden;
-  }
-  
-  .menu{
-    text-align: center;
-  }
-  .open{
-    display: none;
-  }
-  .none{
-    display: inline;
-  }.inline{
-    display: none;
-  }
-}
